# contracts/openapi/agent-public.json
{
  "components": {
    "schemas": {
...
      "BootstrapGovernancePoliciesRequest": {
        "additionalProperties": false,
        "properties": {
          "changeReason": {
            "maxLength": 500,
            "minLength": 10,
            "title": "Changereason",
            "type": "string"
          },
          "expectedExistingCount": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Expectedexistingcount",
            "type": "integer"
          },
          "idempotencyKey": {
            "format": "uuid",
            "title": "Idempotencykey",
            "type": "string"
          }
        },
        "required": [
          "idempotencyKey",
          "expectedExistingCount",
          "changeReason"
        ],
        "title": "BootstrapGovernancePoliciesRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PolicyOutcome": {
        "enum": [
          "ALLOW",
          "HANDOFF",
          "DENY"
        ],
        "title": "PolicyOutcome",
        "type": "string"
      },
...
      "SafetyPolicy": {
        "additionalProperties": false,
        "properties": {
          "maxSourceScopes": {
            "maximum": 7.0,
            "minimum": 1.0,
            "title": "Maxsourcescopes",
            "type": "integer"
          },
          "maxToolCalls": {
            "maximum": 10.0,
            "minimum": 0.0,
            "title": "Maxtoolcalls",
            "type": "integer"
          },
          "mutationOutcome": {
            "$ref": "#/components/schemas/PolicyOutcome"
          },
          "policyVersion": {
            "minimum": 1.0,
            "title": "Policyversion",
            "type": "integer"
          },
          "privilegedDataOutcome": {
            "$ref": "#/components/schemas/PolicyOutcome"
          },
          "promptInjectionOutcome": {
            "$ref": "#/components/schemas/PolicyOutcome"
          },
          "publicWebEnabled": {
            "title": "Publicwebenabled",
            "type": "boolean"
          },
          "requireCitations": {
            "title": "Requirecitations",
            "type": "boolean"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "promptInjectionOutcome",
          "privilegedDataOutcome",
          "mutationOutcome",
          "requireCitations",
          "publicWebEnabled",
          "maxSourceScopes",
          "maxToolCalls",
          "policyVersion",
          "updatedAt"
        ],
        "title": "SafetyPolicy",
        "type": "object"
      },
      "SafetyPolicyEnvelope": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SafetyPolicy"
          },
          "message": {
            "default": "DWAI-ON safety policy loaded.",
            "title": "Message",
            "type": "string"
          },
          "status": {
            "default": "SUCCESS",
            "title": "Status",
            "type": "string"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "data"
        ],
        "title": "SafetyPolicyEnvelope",
        "type": "object"
      },
...
      "UpdateSafetyPolicyRequest": {
        "additionalProperties": false,
        "properties": {
          "changeReason": {
            "maxLength": 500,
            "minLength": 10,
            "title": "Changereason",
            "type": "string"
          },
          "expectedVersion": {
            "minimum": 1.0,
            "title": "Expectedversion",
            "type": "integer"
          },
          "maxSourceScopes": {
            "maximum": 7.0,
            "minimum": 1.0,
            "title": "Maxsourcescopes",
            "type": "integer"
          },
          "maxToolCalls": {
            "maximum": 10.0,
            "minimum": 0.0,
            "title": "Maxtoolcalls",
            "type": "integer"
          },
          "mutationOutcome": {
            "$ref": "#/components/schemas/PolicyOutcome"
          },
          "privilegedDataOutcome": {
            "$ref": "#/components/schemas/PolicyOutcome"
          },
          "requireCitations": {
            "default": true,
            "title": "Requirecitations",
            "type": "boolean"
          }
        },
        "required": [
          "privilegedDataOutcome",
          "mutationOutcome",
          "maxSourceScopes",
          "maxToolCalls",
          "expectedVersion",
          "changeReason"
        ],
        "title": "UpdateSafetyPolicyRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "DWP Agent Runtime",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/admin/safety": {
      "get": {
        "operationId": "get_safety_policy_v1_admin_safety_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-DWP-Tenant-ID",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-Dwp-Tenant-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-DWP-User-ID",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-Dwp-User-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Correlation-ID",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-Correlation-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-DWP-Permissions",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dwp-Permissions"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafetyPolicyEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Safety Policy",
        "tags": [
          "governance"
        ]
      },
      "patch": {
        "operationId": "update_safety_policy_v1_admin_safety_patch",
        "parameters": [
          {
            "in": "header",
            "name": "X-DWP-Tenant-ID",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-Dwp-Tenant-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-DWP-User-ID",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-Dwp-User-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Correlation-ID",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-Correlation-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-DWP-Permissions",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dwp-Permissions"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSafetyPolicyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafetyPolicyEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Safety Policy",
        "tags": [
          "governance"
        ]
      }
    },
    "/v1/admin/safety/bootstrap": {
      "post": {
        "operationId": "bootstrap_safety_policy_v1_admin_safety_bootstrap_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-DWP-Tenant-ID",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-Dwp-Tenant-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-DWP-User-ID",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-Dwp-User-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Correlation-ID",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-Correlation-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-DWP-Permissions",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dwp-Permissions"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BootstrapGovernancePoliciesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafetyPolicyEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bootstrap Safety Policy",
        "tags": [
          "governance"
        ]
      }
    },
...
  }
}